+++ app/src/main/java/de/codevoid/andremote2/PrefKeys.kt
@@ -1,9 +1,10 @@
 package de.codevoid.andremote2
 
 object PrefKeys {
-    const val OVERLAY_SIZE    = "overlay_size"
-    const val OVERLAY_OPACITY = "overlay_opacity"
-    const val OVERLAY_X       = "overlay_x"
-    const val OVERLAY_Y       = "overlay_y"
-    const val JOYSTICK_360    = "joystick_360"
+    const val OVERLAY_SIZE      = "overlay_size"
+    const val OVERLAY_OPACITY   = "overlay_opacity"
+    const val OVERLAY_X         = "overlay_x"
+    const val OVERLAY_Y         = "overlay_y"
+    const val JOYSTICK_360      = "joystick_360"
+    const val OVERLAY_COLLAPSED = "overlay_collapsed"
 }